% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch the PCR simulator Shiny App}
\usage{
run_app()
}
\description{
This function launches the interactive user interface for PCR
for use in validating designed primers bbefore synthesis.
}
